refactor: reorganize metric components with improved structure
- Delete /dev debugging page (no longer needed)
- Extract dialog components for better separation of concerns:
  - create-manual-metric-dialog.tsx (151 lines)
  - edit-metric-dialog.tsx (221 lines)
- Refactor metric-client.tsx → metric-page-client.tsx (300 lines, down from 657)
  - Removed embedded dialog logic
  - Simplified to orchestration and layout only
- Rename and reorganize components:
  - integration-metric-card.tsx → cards/metric-card.tsx
  - dynamic-metric-creator.tsx → forms/template-metric-form.tsx
  - field-renderers/* → forms/fields/*
  - google-sheets-preview.tsx → previews/google-sheets-preview.tsx
- New folder structure:
  _components/
  ├── cards/          # Display components
  ├── dialogs/        # Modal dialogs
  ├── forms/          # Form components
  │   └── fields/     # Reusable form inputs
  ├── previews/       # Integration-specific previews
  └── metric-page-client.tsx

Benefits:
- 40% reduction in main component size (657 → 300 lines)
- Clear separation between display, input, and modal logic
- Better discoverability and maintainability
- Integration-specific code is isolated
